+++ api/src/routes/email.js
@@ -1,0 +1,8 @@
+const { Router } = require("express");
+const router = Router();
+const { getUserByEmailHandler } = require("../handlers/email/handlerGetEmail");
+
+
+router.get("/", getUserByEmailHandler)
+
+module.exports = router;
