+++ client_users/src/pages/account/Account.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import { getUsuarioByID } from "../../redux/actions";
 import Cookies from "js-cookie";
@@ -14,6 +14,7 @@ const Account = () => {
   const decodedToken = jwt_decode(token);
 
   const email = decodedToken.email;
+  console.log(email);
 
   useEffect(() => {
     dispatch(getUsuarioByID(email)); // Actualizar el estado con la respuesta de la acción asincrónica
@@ -23,7 +24,7 @@ const Account = () => {
 
   console.log(userLogin);
 
-  console.log(email);
+ 
 
   return (
     <div className={s.contenedor}>
@@ -50,7 +51,7 @@ const Account = () => {
                     <div className={s.column}>
                       <label>{`${e.primer_nombre} ${e.segundo_nombre} ${e.primer_apellido} ${e.segundo_apellido}`}</label>
                       <label>{e.direccion}</label>
-                      <label>{e.Ciudad.nombre_ciudad}</label>
+                      <label>{e.Ciudad.nombre_ciudad}</label> 
                       <label>{e.telefono}</label>
                       <label>{e.email}</label>
                     </div>
